requirements.txt
Django==2.0.6
djangorestframework==3.8.2
gunicorn==19.9.0
psycopg2==2.7.3.2
psycopg2-binary==2.7.5
Pillow==5.2.0
